+++ awt-metrics/src/lib.rs
@@ -1,5 +1,23 @@
+#![warn(clippy::all)]
+#![warn(clippy::pedantic)]
+#![warn(clippy::cargo)]
+#![allow(unknown_lints)]
+#![warn(missing_debug_implementation)]
+#![warn(missing_copy_implementation)]
+#![warn(rust_2018_idioms)]
+#![warn(rust_2021_compatibility)]
+#![warn(trivial_casts, trivial_numeric_casts)]
+#![warn(unused_qualifications)]
+#![warn(variant_size_difference)]
+
+extern crate alloc;
+
 use core::time::Duration;
 use core::{fmt, fmt::Display, fmt::Formatter};
+
+mod aggregator;
+
+pub use aggregator::Aggregator;
 
 /// Enumerates a metric to trace on a `Request`.
 #[allow(clippy::module_name_repetitions)]